Fix int overflow in MC chunksize tuning

+++ src/enqueue.cc
@@ -1314,11 +1314,12 @@ comp_next:
     work->lastChunkSize = chunkSize / ncclTypeSize(info->datatype);
   } else if (info->algorithm == NCCL_ALGO_MC) {
     if (chunkSize > 131072) chunkSize = 131072;
-    int concurrentOps = info->nChannels*info->comm->channels[0].mc.nHeads;
-    if ((info->nBytes < 512 * (concurrentOps*chunkSize)) && (chunkSize > 65536)) chunkSize = 65536;
-    if ((info->nBytes < 128 * (concurrentOps*chunkSize)) && (chunkSize > 32768)) chunkSize = 32678;
-    if ((info->nBytes < 32 * (concurrentOps*chunkSize)) && (chunkSize > 16384)) chunkSize = 16384;
-    if ((info->nBytes * 4 < (concurrentOps*chunkSize)) && (chunkSize > 8192)) chunkSize = 8192;
+    // Use uint64_t so that concurrentOps*chunkSize*X does not overflow
+    uint64_t concurrentOps = info->nChannels*info->comm->channels[0].mc.nHeads;
+    if ((info->nBytes < (512 * (concurrentOps*chunkSize))) && (chunkSize > 65536)) chunkSize = 65536;
+    if ((info->nBytes < (128 * (concurrentOps*chunkSize))) && (chunkSize > 32768)) chunkSize = 32678;
+    if ((info->nBytes < (32 * (concurrentOps*chunkSize))) && (chunkSize > 16384)) chunkSize = 16384;
+    if (((info->nBytes * 4) < (concurrentOps*chunkSize)) && (chunkSize > 8192)) chunkSize = 8192;
     work->lastChunkSize = chunkSize / ncclTypeSize(info->datatype);
   } else if (info->protocol == NCCL_PROTO_LL) {
     const ssize_t sliceSize = stepSize*sizeof(uint64_t)/sizeof(union ncclLLFifoLine);
